+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.34.4
+### Added
+- Asynchronous preloading only for blocks.
+
 ## 0.34.3
 ### Removed
 - Disable asynchronous preloading to prevent segmentation fault error.